ica_qa_al: Modernize for recent kernels
The `icp_qa_al` module stopped building sometime around
the 3.x series kernels. This commit adds support for this
hardware for the 4.x, 5.x, and 6.x series kernels. Outside
of support for interface changes, Osal (OS abstraction
layer) interfaces are modified to begin supporting the
Y2038 problem (note that overflows may still be present in
the code at this time).

New versions of Clang also picked up a bug via static
analysis wherein a wrong enum value was being passed in
some crypto code (eek), so this commit also fixes that.

+++ icp_qa_al/osal/include/OsalTypes.h
@@ -410,7 +410,7 @@ typedef void (*OsalVoidFnPtr) (void);
  */
 typedef struct
 {
-    UINT32 secs;        /**< seconds */
+    UINT64 secs;        /**< seconds */
     UINT32 nsecs;        /**< nanoseconds */
 } OsalTimeval;
